// packages/types/src/generated/ContourSnapshot.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ContourId } from "../ids";

import type { PointSnapshot } from "./PointSnapshot";

export type ContourSnapshot = { id: ContourId, points: Array<PointSnapshot>, closed: boolean, };