// packages/types/src/generated/GlyphSnapshot.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ContourId } from "../ids";

import type { AnchorSnapshot } from "./AnchorSnapshot";
import type { ContourSnapshot } from "./ContourSnapshot";
import type { RenderContourSnapshot } from "./RenderContourSnapshot";

export type GlyphSnapshot = { unicode: number, name: string, xAdvance: number, contours: Array<ContourSnapshot>, anchors: Array<AnchorSnapshot>, compositeContours: Array<RenderContourSnapshot>, activeContourId: ContourId | null, };